refactor(product): extract and generalize ProductCardList for reuse
