--- src/pages/HomePage/ui/ProductCardList/index.ts
@@ -1,1 +1,0 @@
-export { ProductCardList } from './ProductCardList';
